Add new transmission request for the Stream Scheduler
Due to the new scheduler scheduling by stream, not by mobile station, a
new transmission request is needed. This request has src and destination
attributes, identifying the stream as well as the packets currently
queued for that stream at the sender, which might be either the
originating mobile station or the Base station forwarding the package to
the destination.