Suppress deprecated warnings from HORIZON
Horizon uses the cLinkedList class extensively, which seems to be
deprecated. With -Wall compiler parameter, this leads to very cluttered
compile outputs, showing a lot of deprecated warnings the user can't do
anything about, and obfuscating actual warnings from the model code.

Deprecated warnings are now completely suppressed via the "includes.h"
file.

For now, this does not work for the automatically generated message
source files. But, as they are automatically generated an not written by
the user, warnings from them can be safely ignored.