Add StreamScheduler class for stream to RB assignment
This is the main scheduler module, tasked with assigning packet streams
to ressource blocks. The actual scheduling of the packets, aka the
decision which packets will be send in any given TTI is done in the
individual resource block schedulers.

The StreamScheduler module implementation provides only a simple Round
Robin assignment of streams to resource blocks. In general, new
scheduling algorithms only need to reimplement the scheduleStreams
method.