Fix multiple memory errors in METIS calc when there are no neighbours
Many of the arrays used in the METIS:init calculations are sized with
the number of interferers, but later they are actually written to for
the current BS too, leading to invalid writes to arbitrary memory
positions. This problem occurs when neighbourPositions only contains one
BS, in single cell simulations. Then, numOfInterferers is 0. If
numOfInterferers is now used as the size for arrays, a 0-length array is
created and no memory is allocated. But, those same array dimensions are
later written to, leading to a large number of instances for undefined
behaviour, random segfaults and incorrect results.