Fix/Add destruction of BsChannel/METIS
The single METIS channel instance used in the simulation and shared
amongst all BSChannels needs to be destructed somewhere, but only once,
not in each instance destructor of each BsChannel instance. Thus, it is
now destructed only by the channel with the index 0, which is also the
one sending around the initial pointer message.

This is still an ungly solution, the entire handing around of the
channel pointer via messages and the fact that each BSChannel instance
creates a METIS channel, only to replace it with the one handed around
via the POINTER_EXCHANGE2 Msg. Perhaps some kind of Singleton is in
order, which ought to be possible considering that the METISChannel
class does not subclass any OMNET classes.