Move position dependent METIS computations to own method
Many METIS channel computations depend on the position of the mobile
stations, which might change. Repeatedly calling the init method was not
a good design and the recomputeMETISParams method can now be called to
recompute all METIS values.

This also fixes all memory leaks and allows for repeated computations
with different mobile positions.