Add method to get a neighbours number of mobile stations
To compute the interference part of the SINR equation, each METIS
instance needs the number of and positions for all neighbouring cell's
mobile stations. While position messages of all mobile stations are
already forwarded to all neighbouring cells, there was now way of
accessing the number of MS. This is now accomplished through the
NeighbourIdMatching class which has access to neighbouring cell's NED
parameters, the number of mobile stations among them.

This implementation will need to be changed to something more dynamic if
the number of mobile stations ever changes in the course of a simulation
run.