Add TransInfoBs message type
This new type of message is send by all Base Stations to all Mobile
Stations outside their own cells. It contains the transmission power as
well as the used resource block. This information is necessary for the
Mobile Station channel module to compute the SINR interference values
when receiving messages form their own Base stations. The channel module
needs to know which other Base Stations are sending on the resource
block in the current TTI.