Extend LSP grid to entire playground
Previously, only the current cell was overlaid with the LSP grid for
correlation. But, with the multi cell simulations, we need correlation
between MS/BS in multiple different cells, hence the expansion of the
LSP computation grid from a single cell to the entire playground.

This consumes quite some time.