Activate random MS placement
Up to now, all Mobile Stations were position at the beginning of a "road"
and moved linearly along that road in the course of the simulation.

With this commit, the MS placement algorithm can be choosen via the
"positioning" ini parameter. In the MsMac::init method, the inital
position is then computed by the algorithm determined by that method.
All available choices can be found in MsMac::Placement enum.