Add output of coefficient tables to METIS channel
Whenever the METIS channel receives a message with the "DEBUG" name, it
outputs the Up/Down tables into files in long format tables. The DEBUG
message is issued by the MsMac module in every cell whenever the NDEBUG
preprocessor variable is not set. (So it won't be send when the
simulator is compiled in Release mode)