Fix index errors in MS Positions array
In the array holding all MS positions sorted by their cell's BS id, the
actual BS id was not used correctly when building the array for base
stations with an id>0. This lead to problems in the METIS calculations,
because there it is assumed that the index of the first dimension of the
MS positions array corresponds to the BS id.