Read communication streams from dedicated XML file
Instead of reading stream definitions from a string given in Omnet ini
parameters, they are now read from a XML file. The XML format is
intended to define all streams for all cells and MS in a single file.

This allows for easier generation of larger communication tables with
any number of XML manipulation tools ans also precludes the need for
involved parsers for a custom file format.