Distinguish between Uplink/Downlink resource blocks
Up to now, Resource blocks were assumed to be balanced - there were as
many Up- as down blocks. This with the possiblity to have different
numbers of RBs for the different link directions, scheduling needed an
update. Now, there is not a single RBScheduler per resource block, but
one per direction per resource block.

To allow for scheduling, stream transmission requests now provide
the transmission direction they are concerned with. This information
also replaces the "from BS" attribute of TransRequest messages.