Add FDD to METIS channel calculations
FDD (Frequency Division Duplex) denotes the technique to divide a given
frequency band in half, using one half (in this implementation the upper
one) for the
uplink transmissions from mobile stations to base stations and the lower
half for the downlink from base stations to mobile stations. With this
technique, a cell does not need different phases for up- and downlink.
Instead, communications in both directions can occur in the same time
slot.