Generalize Transmission information on direction, not origin
With the advent of D2D communication, possible interferers can not be
determined by who is sending. This stems from the fact that D2D
transmissions get dynamically assigned to the UP or DOWN half of the
transmission band. To keep up with who can interfere with whom,
TransInfo messages need information on the transmission direction.

Up/Down stay as they are. But D2D transmissions need additional
information, namely that they are D2D, so all parts of the program know
where to take the link values from.

This commit mostly introduces a generalized TransInfo message which is
now routed by the direction it's originator is sending on, not on
whether that originator is a MS or BS.