Add handling of D2D streams to schedulers
Whenever a D2D stream is to be scheduled, it is assigned first to either
the UP link or DOWN link portion of the frequency band and then assigned
a specific resource block from that portion.

Schedule requests and Schedules themselves now also contain the
information on whether the scheduled packet is D2D or not.