Remove minimum distance requirements for pathloss calculations
Fully random positioning of Mobile Stations in combination with D2D link
introduction leads to a lot of links with distances below the minimum of
10 meters, as defined in METIS D1.2.

Instead of introducing complex positioning algorithms, new formulas will
be used for pathloss calculations in the future. Until then, we simply
remove the distance checks completely.