Reintroduce SINR estimates
SINR estimates are now conducted by computing the average SINR value
over all possible receivers for a given resource block and sender. They
are computed at the and of a TTI for the next TTI, taking the
interference from the previous TTI into account.

This information is stored in the BsMac/MsMac to be used in computing
the amount of data which can be send in a TTI over any given RB and it
is also forwarded to the scheduler to be used as a quality measure when
assigning resource blocks to MS and BS.