Adapt transmission info distribution to multiple packets
The generation of transmission information messages needed to be adapted
for multiple packets per TTI simulation. Before, one TransInfo message
was send for each packet transmitted. But, this becomes wrong as soon as
multiple packets are send in the same TTI using the same RB. With the
old scheme, the interference by those multiple packets would all be
added to the overall interference on that RB. This would lead to an
underestimate for the overall SINR value.

Now, only one TransInfo message is send by each MS per RB used.