Move TransInfo handling into Channel
TransInfo messages for interference computations have been stored in the
MsChannel/BsChannel classes. This architecture lead to waste in
compuational as well as memory resources, as each MS and BS has multiple
channel instances. But the TransInfo messages are only needed in the
Channel implementation when computing interference, and there is only
one channel instance per cell. Thus, it made sense to store these
messages in the channel itself.