Add implementation of table based rate/capacity computation
The KoI project has done measurements for different combinations of
modulation and coding schemes as well as resource block bandwidth. These
measurements will now be used, read from a tabular file, by the Coding
class to determine the number of bits transmittable on a resource block
with a given SINR value.