Work around broken HORIZON XML caching
For some reason, HORIZON runs into problems when too many modules (in
this case, Mobile Stations) want to access the same XML file. At some
point, it just seems to remove the file from the cache, and then throws
an error when the next MS wants to initialize itself with the stream
table.

My best guess is that something is wrong with the caching. But I don't
have the time to fix it, so I just store the root node in a static
variable on the traffic generator, which seems to solve the problem.

Now, there is only one call to the getXMLDocument method, by the very
first MS which is getting initialized. This MS then stores a pointer to
the parsed XML file, which is subsequently used by the other mobile
stations, because they are all using the same XML table.