Add templated type for multidimensional vectors
To avoid memory leaks and make use of more modern idioms, several
multidimensional arrays in the channel implementations had been replaced
with vectors. The only drawback of such an approach is the rather ugly
syntax for vectors of higher dimensionality.

To alleviate that problem, the VectorNd type has been introduced. This
type allows the definition of vectors with arbitrary dimensions with a
simple template.

This commit reduces vector<vector<vector<double>>> to
VectorNd<double,3>.