Fix Radians/Degrees mixup
The Zenith/Azimuth angles generated are given in Degrees, so they don't
need to be converted in the Ray Sum computations. But the LOS Rays are
computed with the LOS direction angles, which are computed in radias and
thus need to be converted to Degrees in the final Ray computations.