Add time factor MCS SINR thresholds
The thresholds for using a specific MCS are now modified by the ratio of
the reference TTI to the actual TTI. THis means, if the actual TTI is
longer, a given MCS can be used with worse SINR values. If it is shorter
in relation to the reference TTI, an MCS needs a better SINR to be used.